Document new behaviour of Docker on Mac
The MySQL container seems to take longer to start on Docker for Mac or the Rails container does not wait for
MySQL to start. It is now documented that the user has to start the db container first before running the
database setup.
Also I droped the db data container as it didn't work as expected. The MySQL data is now stored in a directory
on the host. Maybe I'll try that again sometimes.